LayoutTests/imported/w3c/web-platform-tests/shadow-dom/declarative/declarative-shadow-dom-slot-assignment-serialization.tentative-expected.txt


FAIL shadowrootslotassignment=manual is serialized and appears before shadowrootclonable and shadowrootserializable assert_equals: expected "manual" but got "named"
PASS shadowrootslotassignment=named is not serialized as it's the default
FAIL shadowrootslotassignment=manual serializes between shadowrootdelegatesfocus and shadowrootclonable assert_equals: expected "manual" but got "named"
